+++ packages/timeline-state-resolver-types/src/device.ts
@@ -20,6 +20,20 @@ import {
 } from '.'
 import { ShotokuOptions } from './shotoku'
 
+export enum StatusCode {
+	UNKNOWN = 0, // Status unknown
+	GOOD = 1, // All good and green
+	WARNING_MINOR = 2, // Everything is not OK, operation is not affected
+	WARNING_MAJOR = 3, // Everything is not OK, operation might be affected
+	BAD = 4, // Operation affected, possible to recover
+	FATAL = 5, // Operation affected, not possible to recover without manual interference
+}
+export interface DeviceStatus {
+	statusCode: StatusCode
+	messages: Array<string>
+	active: boolean
+}
+
 export interface DeviceOptionsBase<T> extends SlowReportOptions {
 	type: DeviceType
 	isMultiThreaded?: boolean